add question detail html

+++ src/main/java/com/community/community/repository/QuestionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 // This will be AUTO IMPLEMENTED by Spring into a Bean called QuestionRepository
 // CRUD refers Create, Read, Update, Delete
-public interface QuestionRepository extends CrudRepository<Question, Integer> {
+public interface QuestionRepository extends CrudRepository<Question, Long> {
 
 }
